refactor(user_properties) rename package matching java convention
