+++ src/tool/clients/forms/FormsClient.java
@@ -24,7 +24,7 @@ import javafx.scene.layout.BorderPane;
 import javafx.stage.Stage;
 import tool.clients.Client;
 import tool.clients.EventHandler;
-import tool.helper.userProperties.PropertyManager;
+import tool.helper.user_properties.PropertyManager;
 import tool.xmodeler.XModeler;
 import xos.Message;
 import xos.Value;
