+++ src/tool/xmodeler/XModeler.java
@@ -31,8 +31,8 @@ import tool.clients.workbench.WorkbenchClient;
 import tool.console.Console;
 import tool.console.ConsoleClient;
 import tool.helper.IconGenerator;
-import tool.helper.userProperties.PropertyManager;
-import tool.helper.userProperties.UserProperty;
+import tool.helper.user_properties.PropertyManager;
+import tool.helper.user_properties.UserProperty;
 import xos.OperatingSystem;
 
 import javax.xml.parsers.DocumentBuilder;
